Use external 74HC595 and read value of Q0-Q2 on STCP
The frequency of SHCP is around 180kHz, which is probably
too fast to handle in the CPU. Instead I have added an
external 74HC595 which handles the high frequency part.
The STCP has a frequency of about 190Hz which should be
easier to handle in SW. Also disable logging from interrupt
with a special compile flag.