Add support for INH-pin and make most buttons work
I forgot to use the INH pin value (Q3 on the shift register).
It also seems like the right side of the control card has the
exact same connections but with inverted INH pin. With this
commit all buttons but the power on works. It might be that
the power on button has an extra safety thing that prevents
it from working. This can be figured out at a later stage, but
perhaps it is just good that it is not possible to turn on from
the ioio.