Remove some unused interfaces and add valid check to temperature
The getEvent mechanism were never used, lets get rid of them.
Also make sure that the temperature is in reasonable range.
There should also be a check for the checksum (if any), but
I have not yet checked how that works. We should perhaps
also see if we don't get any new readings within some time
and then notify the temperature as invalid.